Bug 1802584 - Announce "Find in page" edit box when using Talkback
When using talkback, the focus keeps jumping between the dismiss
button and the "Find in page" edit text. Seems like giving up
the view stub fixes this issue. Considering that the view inflation
performance wins are rather small for such added complexity,
we have given up the view stub.
