--- fenix/app/src/main/res/layout/stub_find_in_page.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- This Source Code Form is subject to the terms of the Mozilla Public
-   - License, v. 2.0. If a copy of the MPL was not distributed with this
-   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
-<mozilla.components.feature.findinpage.view.FindInPageBar xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/findInPageView"
-    android:layout_width="match_parent"
-    android:layout_height="56dp"
-    android:layout_gravity="bottom"
-    android:background="?attr/layer1"
-    android:clickable="true"
-    android:focusable="true"
-    app:findInPageNoMatchesTextColor="?attr/textWarning"
-    app:findInPageButtonsTint="?attr/textPrimary"
-    app:findInPageResultCountTextColor="?attr/textPrimary" />
